requirements.txt
Flask>=2.3
reportlab>=4.0
python-dotenv>=1.0
twilio>=9.0
scikit-learn>=1.3
joblib>=1.3
numpy>=1.26
